+++ src/Classes/FinalRule.php
@@ -16,23 +16,40 @@ namespace Localheinz\PHPStan\Rules\Classes;
 use PhpParser\Node;
 use PHPStan\Analyser\Scope;
 use PHPStan\Rules\Rule;
+use PHPStan\ShouldNotHappenException;
 
 final class FinalRule implements Rule
 {
+    /**
+     * @var string[]
+     */
+    private $excludedClassNames;
+
+    /**
+     * @param string[] $excludedClassNames
+     */
+    public function __construct(array $excludedClassNames = [])
+    {
+        $this->excludedClassNames = \array_map(static function (string $excludedClassName): string {
+            return $excludedClassName;
+        }, $excludedClassNames);
+    }
+
     public function getNodeType(): string
     {
         return Node\Stmt\Class_::class;
     }
 
-    /**
-     * @param Node\Stmt\Class_ $node
-     * @param Scope            $scope
-     *
-     * @return string[] errors
-     */
     public function processNode(Node $node, Scope $scope): array
     {
-        if (!isset($node->namespacedName) || $node->isFinal()) {
+        if (!$node instanceof Node\Stmt\Class_) {
+            throw new ShouldNotHappenException();
+        }
+
+        if (!isset($node->namespacedName)
+            || \in_array($node->namespacedName->toString(), $this->excludedClassNames, true)
+            || $node->isFinal()
+        ) {
             return [];
         }